+++ src/de/lmu/dbs/ciaa/classifier/features/FeatureHarmonic3.java
@@ -47,7 +47,7 @@ public class FeatureHarmonic3 extends Feature {
 		}
 		this.uY = (int)(params.binsPerOctave * harmonics[RandomUtils.randomInt(harmonics.length-1)]);
 		this.vY = (int)(params.binsPerOctave * harmonics[RandomUtils.randomInt(harmonics.length-1)]);
-		this.threshold = RandomUtils.randomInt(getMaxValue());
+		this.threshold = (int)RandomUtils.randomInt((int)getMaxValue());
 		//generateHarmonicFactors(50);
 	}
 	
@@ -61,7 +61,7 @@ public class FeatureHarmonic3 extends Feature {
 		this.uY = f.uY;
 		this.vX = f.vX;
 		this.vY = f.vY;
-		this.threshold = RandomUtils.randomInt(getMaxValue());
+		this.threshold = (int)RandomUtils.randomInt((int)getMaxValue());
 	}
 	
 	/**
@@ -98,7 +98,7 @@ public class FeatureHarmonic3 extends Feature {
 	 * @return
 	 * @throws Exception 
 	 */
-	public int evaluate(final byte[][] data, final int x, final int y) throws Exception {
+	public float evaluate(final byte[][] data, final int x, final int y) throws Exception {
 		// If out of range on time axis, the evaluation is useless, so return a minimal value
 		int iuX = x+uX;
 		if (iuX < 0 || iuX >= data.length) return Integer.MIN_VALUE;
@@ -119,7 +119,7 @@ public class FeatureHarmonic3 extends Feature {
 	}
 	
 	@Override
-	public int getMaxValue() {
+	public float getMaxValue() {
 		return (Byte.MAX_VALUE-1) * 3;
 	}
