+++ src/de/lmu/dbs/ciaa/classifier/features/FeatureHarmonic4.java
@@ -49,7 +49,7 @@ public class FeatureHarmonic4 extends Feature {
 		this.vY = (int)(params.binsPerOctave * harmonics[RandomUtils.randomInt(harmonics.length-1)]);
 		while(this.vY == this.uY)
 			this.vY = (int)(params.binsPerOctave * harmonics[RandomUtils.randomInt(harmonics.length-1)]);
-		this.threshold = RandomUtils.randomInt(getMaxValue());
+		this.threshold = (int)RandomUtils.randomInt((int)getMaxValue());
 		//generateHarmonicFactors(50);
 	}
 	
@@ -63,7 +63,7 @@ public class FeatureHarmonic4 extends Feature {
 		this.uY = f.uY;
 		this.vX = f.vX;
 		this.vY = f.vY;
-		this.threshold = RandomUtils.randomInt(getMaxValue());
+		this.threshold = (int)RandomUtils.randomInt((int)getMaxValue());
 	}
 	
 	/**
@@ -100,7 +100,7 @@ public class FeatureHarmonic4 extends Feature {
 	 * @return
 	 * @throws Exception 
 	 */
-	public int evaluate(final byte[][] data, final int x, final int y) throws Exception {
+	public float evaluate(final byte[][] data, final int x, final int y) throws Exception {
 		// If out of range on time axis, the evaluation is useless, so return a minimal value
 		int iuX = x+uX;
 		if (iuX < 0 || iuX >= data.length) return Integer.MIN_VALUE;
@@ -121,7 +121,7 @@ public class FeatureHarmonic4 extends Feature {
 	}
 	
 	@Override
-	public int getMaxValue() {
+	public float getMaxValue() {
 		return (Byte.MAX_VALUE-1)*(Byte.MAX_VALUE-1) * 4;
 	}
